feat: 更新到v1.2.2版本 - 学习分析功能增强和文档完善
主要更新内容：
- 修复学习分析数据类型错误
- 更新所有install目录下的文档到v1.2.2版本
- 完善API接口文档，补充学习分析相关接口
- 更新功能说明文档，详细描述学习分析功能
- 优化快速体验指南，添加新功能体验步骤
- 完善部署指南，添加依赖更新和故障排除
- 前端AI助手功能优化
- 新增PPT模板选择器组件

+++ backend/routes/advanced_analytics.py
@@ -388,7 +388,7 @@ def _calculate_learning_health_score(efficiency_data, time_data, pattern_data):
     if time_data and time_data.get('weekly_distribution'):
         weekly_dist = time_data['weekly_distribution']
         active_days = sum(1 for day in weekly_dist if day.get('question_count', 0) > 0)
-        scores['balance_score'] = (active_days / 7) * 100
+        scores['balance_score'] = int((active_days / 7) * 100)
     
     # 习惯评分
     if pattern_data and pattern_data.get('habit_assessment'):
